fix: prospector reports
